refactor: extract shared GT API adapter into core
