+++ .changeset/shared-api-adapter.md
@@ -1,0 +1,7 @@
+---
+'generaltranslation': minor
+'gt': patch
+'gt-sanity': patch
+---
+
+Extract the shared generated-API adapter into core so CLI and Sanity use one locale canonicalization and response-mapping boundary.
