+++ packages/cli/src/utils/api.ts
@@ -1,19 +1,10 @@
-import {
-  resolveAliasLocale,
-  resolveCanonicalLocale,
-} from '@generaltranslation/format';
+import { resolveCanonicalLocale } from '@generaltranslation/format';
 import type { CustomMapping } from '@generaltranslation/format/types';
 import {
   createApiClient,
-  createBranch,
   createProject,
   createTag,
-  DEFAULT_BATCH_SIZE,
-  downloadFiles,
-  enqueueFileTranslations,
-  generateProjectContext,
   getBranchInfo,
-  getFileInfo,
   getOrphanedFiles,
   getProjectInfo,
   getTranslationJobInfo,
@@ -22,62 +13,33 @@ import {
   processFileMoves,
   publishFiles,
   submitUserEditDiffs,
-  uploadSourceFiles,
-  uploadTranslations,
   type ApiClientConfig,
   type AwaitJobsOptions,
-  type CreateBranchData,
   type CreateProjectData,
   type CreateTagData,
-  type DownloadFilesData,
-  type EnqueueFileTranslationsData,
-  type GenerateProjectContextData,
   type GetBranchInfoData,
-  type GetFileInfoData,
   type ProcessFileMovesData,
   type PublishFilesData,
   type SubmitUserEditDiffsData,
-  type UploadSourceFilesData,
-  type UploadTranslationsData,
 } from 'generaltranslation/api';
 import {
-  decodeFileContent,
-  defaultBaseUrl,
-  encodeFileContent,
+  createGtApiAdapter,
   unwrapApiResult,
 } from 'generaltranslation/internal';
 
-let client: ReturnType<typeof createApiClient> | undefined;
-let configuredClientConfig: ApiClientConfig | undefined;
+const {
+  configure: configureSharedApi,
+  getClient,
+  getClientConfig,
+  ...sharedApi
+} = createGtApiAdapter();
 let customMapping: CustomMapping | undefined;
-
-function getClient(): ReturnType<typeof createApiClient> {
-  if (!client) {
-    throw new Error(
-      'API client not configured — call configureApiClient first'
-    );
-  }
-  return client;
-}
-
-function getClientConfig(): ApiClientConfig {
-  if (!configuredClientConfig) {
-    throw new Error(
-      'API client not configured — call configureApiClient first'
-    );
-  }
-  return configuredClientConfig;
-}
 
 export function configureApiClient(
   config: ApiClientConfig & { customMapping?: CustomMapping }
 ): void {
-  const { customMapping: mapping, ...clientConfig } = config;
-  client = createApiClient(clientConfig);
-  // Retain the resolved config for the one project-info call that deliberately
-  // uses a shorter timeout than normal translation requests.
-  configuredClientConfig = clientConfig;
-  customMapping = mapping;
+  configureSharedApi(config);
+  customMapping = config.customMapping;
 }
 
 export function createNonRetryingApiClient(config: ApiClientConfig) {
@@ -85,41 +47,9 @@ export function createNonRetryingApiClient(config: ApiClientConfig) {
 }
 
 export const api = {
+  ...sharedApi,
   async queryBranchData(body: GetBranchInfoData['body']) {
     return unwrapApiResult(await getBranchInfo({ body, client: getClient() }));
-  },
-
-  async createBranch(body: CreateBranchData['body']) {
-    return unwrapApiResult(await createBranch({ body, client: getClient() }));
-  },
-
-  async queryFileData(body: GetFileInfoData['body']) {
-    const result = unwrapApiResult(
-      await getFileInfo({
-        body: {
-          ...body,
-          translatedFiles: body.translatedFiles?.map((file) => ({
-            ...file,
-            locale: resolveCanonicalLocale(file.locale, customMapping),
-          })),
-        },
-        client: getClient(),
-      })
-    );
-    return {
-      ...result,
-      translatedFiles: result.translatedFiles.map((file) => ({
-        ...file,
-        locale: resolveAliasLocale(file.locale, customMapping),
-      })),
-      sourceFiles: result.sourceFiles.map((file) => ({
-        ...file,
-        sourceLocale: resolveAliasLocale(file.sourceLocale, customMapping),
-        locales: file.locales.map((locale) =>
-          resolveAliasLocale(locale, customMapping)
-        ),
-      })),
-    };
   },
 
   async checkJobStatus(jobIds: string[]) {
@@ -145,43 +75,6 @@ export const api = {
         path: { projectId: config.projectId },
       })
     );
-  },
-
-  resolveAliasLocale(locale: string) {
-    return resolveAliasLocale(locale, customMapping);
-  },
-
-  async downloadFileBatch(files: DownloadFilesData['body']) {
-    if (files.length === 0) return { files: [], count: 0, pending: [] };
-
-    const request = async (batch: DownloadFilesData['body']) =>
-      unwrapApiResult(
-        await downloadFiles({
-          body: batch.map((file) => ({
-            ...file,
-            locale: file.locale
-              ? resolveCanonicalLocale(file.locale, customMapping)
-              : undefined,
-          })),
-          client: getClient(),
-        })
-      );
-    const responses = await processBatches(files, async (batch) => [
-      await request(batch),
-    ]);
-    return {
-      files: responses.flatMap((response) =>
-        response.files.map((file) => ({
-          ...file,
-          ...(file.locale && {
-            locale: resolveAliasLocale(file.locale, customMapping),
-          }),
-          data: decodeFileContent(file.data, file.fileFormat),
-        }))
-      ),
-      count: responses.reduce((count, response) => count + response.count, 0),
-      pending: responses.flatMap((response) => response.pending ?? []),
-    };
   },
 
   async publishFiles(files: PublishFilesData['body']['files']) {
@@ -281,28 +174,6 @@ export const api = {
     };
   },
 
-  async setupProject(
-    files: GenerateProjectContextData['body']['files'],
-    options: Omit<GenerateProjectContextData['body'], 'files'> = {}
-  ) {
-    return unwrapApiResult(
-      await generateProjectContext({
-        body: {
-          files: files.map(({ branchId, fileId, versionId }) => ({
-            branchId,
-            fileId,
-            versionId,
-          })),
-          locales: options.locales?.map((locale) =>
-            resolveCanonicalLocale(locale, customMapping)
-          ),
-          force: options.force,
-        },
-        client: getClient(),
-      })
-    );
-  },
-
   async awaitJobs(jobIds: readonly string[], options?: AwaitJobsOptions) {
     // Poll through unwrapApiResult so HTTP failures throw ApiError with the
     // status code instead of the decoded response body.
@@ -318,126 +189,6 @@ export const api = {
         ),
       options
     );
-  },
-
-  async enqueueFiles(
-    files: EnqueueFileTranslationsData['body']['files'],
-    options: {
-      sourceLocale?: string;
-      targetLocales: string[];
-      modelProvider?: string;
-      force?: boolean;
-    }
-  ) {
-    const targetLocales = options.targetLocales.map((locale) =>
-      resolveCanonicalLocale(locale, customMapping)
-    );
-    const result = await processBatches(files, async (batch) => {
-      const response = unwrapApiResult(
-        await enqueueFileTranslations({
-          body: {
-            files: batch.map(
-              ({ branchId, fileId, versionId, fileName, transformFormat }) => ({
-                branchId,
-                fileId,
-                versionId,
-                fileName,
-                transformFormat,
-              })
-            ),
-            targetLocales,
-            sourceLocale: options.sourceLocale
-              ? resolveCanonicalLocale(options.sourceLocale, customMapping)
-              : undefined,
-            // The CLI intentionally accepts custom model-provider strings beyond
-            // the OpenAPI ANTHROPIC|OPENAI|XAI|GOOGLE enum; preserve wire behavior.
-            modelProvider:
-              options.modelProvider as EnqueueFileTranslationsData['body']['modelProvider'],
-            force: options.force,
-          },
-          client: getClient(),
-        })
-      );
-      const jobData = 'jobData' in response ? response.jobData : response.data;
-      return Object.entries(jobData);
-    });
-
-    return {
-      jobData: Object.fromEntries(result),
-      locales: targetLocales,
-      message: `Successfully enqueued ${result.length} file translation jobs in ${Math.ceil(files.length / DEFAULT_BATCH_SIZE)} batch(es)`,
-    };
-  },
-
-  async uploadSourceFiles(
-    files: Array<{
-      source: UploadSourceFilesData['body']['data'][number]['source'];
-    }>,
-    options: { sourceLocale: string }
-  ) {
-    const sourceLocale = resolveCanonicalLocale(
-      options.sourceLocale,
-      customMapping
-    );
-    const result = await processBatches(files, async (batch) => {
-      const response = unwrapApiResult(
-        await uploadSourceFiles({
-          body: {
-            data: batch.map(({ source }) => ({
-              source: {
-                ...source,
-                content: encodeFileContent(source.content, source.fileFormat),
-                locale: resolveCanonicalLocale(source.locale, customMapping),
-              },
-            })),
-            sourceLocale,
-          },
-          client: getClient(),
-        })
-      );
-      return response.uploadedFiles;
-    });
-
-    return { uploadedFiles: result };
-  },
-
-  async uploadTranslations(
-    files: UploadTranslationsData['body']['data'],
-    options: { sourceLocale: string }
-  ) {
-    const result = await processBatches(files, async (batch) => {
-      const response = unwrapApiResult(
-        await uploadTranslations({
-          body: {
-            data: batch.map(({ source, translations }) => ({
-              source: {
-                ...source,
-                content: encodeFileContent(source.content, source.fileFormat),
-              },
-              translations: translations.map((translation) => ({
-                ...translation,
-                locale: resolveCanonicalLocale(
-                  translation.locale,
-                  customMapping
-                ),
-                content: encodeFileContent(
-                  translation.content,
-                  translation.fileFormat
-                ),
-              })),
-            })),
-            sourceLocale: resolveCanonicalLocale(
-              options.sourceLocale,
-              customMapping
-            ),
-          },
-          client: getClient(),
-        })
-      );
-      return response.uploadedFiles;
-    });
-
-    return { uploadedFiles: result };
   },
 };
