+++ packages/sanity/src/adapter/api.ts
@@ -1,36 +1,23 @@
 import {
   awaitJobs as awaitApiJobs,
-  createApiClient,
-  createBranch,
   downloadFile,
-  downloadFiles,
-  enqueueFileTranslations,
-  generateProjectContext,
-  getFileInfo,
   getTranslationStatus,
-  processBatches,
-  uploadSourceFiles,
-  uploadTranslations,
   type ApiClientConfig,
-  type CreateBranchData,
-  type DownloadFilesData,
-  type EnqueueFileTranslationsData,
-  type GenerateProjectContextData,
-  type GetFileInfoData,
-  type UploadSourceFilesData,
-  type UploadTranslationsData,
 } from 'generaltranslation/api';
 import { resolveAliasLocale, resolveCanonicalLocale } from 'generaltranslation';
 import {
+  createGtApiAdapter,
   decode as decodeBase64,
-  decodeFileContent,
   defaultBaseUrl,
-  encodeFileContent,
   unwrapApiResult,
 } from 'generaltranslation/internal';
-import type { CustomMapping, DownloadedFile } from 'generaltranslation/types';
+import type { CustomMapping } from 'generaltranslation/types';
 
-let client = createApiClient({ baseUrl: defaultBaseUrl });
+const {
+  configure: configureSharedApi,
+  getClient,
+  ...sharedApi
+} = createGtApiAdapter({ baseUrl: defaultBaseUrl });
 let customMapping: CustomMapping | undefined;
 
 export function configureApiClient(
@@ -39,116 +26,12 @@ export function configureApiClient(
     customMapping?: CustomMapping;
   }
 ): void {
-  const { customMapping: mapping, ...clientConfig } = config;
-  client = createApiClient({
-    baseUrl: defaultBaseUrl,
-    ...clientConfig,
-  });
-  customMapping = mapping;
+  configureSharedApi({ baseUrl: defaultBaseUrl, ...config });
+  customMapping = config.customMapping;
 }
 
 export const api = {
-  async uploadSourceFiles(
-    files: Array<{
-      source: UploadSourceFilesData['body']['data'][number]['source'];
-    }>,
-    options: { sourceLocale: string }
-  ) {
-    const sourceLocale = resolveCanonicalLocale(
-      options.sourceLocale,
-      customMapping
-    );
-    const result = await processBatches(files, async (batch) => {
-      const response = unwrapApiResult(
-        await uploadSourceFiles({
-          body: {
-            data: batch.map(({ source }) => ({
-              source: {
-                ...source,
-                content: encodeFileContent(source.content, source.fileFormat),
-                locale: resolveCanonicalLocale(source.locale, customMapping),
-              },
-            })),
-            sourceLocale,
-          },
-          client,
-        })
-      );
-      return response.uploadedFiles;
-    });
-    return { uploadedFiles: result };
-  },
-
-  async uploadTranslations(
-    files: UploadTranslationsData['body']['data'],
-    options: { sourceLocale: string }
-  ) {
-    const result = await processBatches(files, async (batch) => {
-      const response = unwrapApiResult(
-        await uploadTranslations({
-          body: {
-            data: batch.map(({ source, translations }) => ({
-              source: {
-                ...source,
-                content: encodeFileContent(source.content, source.fileFormat),
-              },
-              translations: translations.map((translation) => ({
-                ...translation,
-                content: encodeFileContent(
-                  translation.content,
-                  translation.fileFormat
-                ),
-                locale: resolveCanonicalLocale(
-                  translation.locale,
-                  customMapping
-                ),
-              })),
-            })),
-            sourceLocale: resolveCanonicalLocale(
-              options.sourceLocale,
-              customMapping
-            ),
-          },
-          client,
-        })
-      );
-      return response.uploadedFiles;
-    });
-    return { uploadedFiles: result };
-  },
-
-  async enqueueFiles(
-    files: EnqueueFileTranslationsData['body']['files'],
-    options: {
-      sourceLocale?: string;
-      targetLocales: string[];
-      force?: boolean;
-    }
-  ) {
-    const sourceLocale = options.sourceLocale
-      ? resolveCanonicalLocale(options.sourceLocale, customMapping)
-      : undefined;
-    const targetLocales = options.targetLocales.map((locale) =>
-      resolveCanonicalLocale(locale, customMapping)
-    );
-    const result = await processBatches(files, async (batch) => {
-      const response = unwrapApiResult(
-        await enqueueFileTranslations({
-          body: {
-            files: batch,
-            sourceLocale,
-            targetLocales,
-            force: options.force,
-          },
-          client,
-        })
-      );
-      return Object.entries(
-        'jobData' in response ? response.jobData : response.data
-      );
-    });
-    return { jobData: Object.fromEntries(result), locales: targetLocales };
-  },
+  ...sharedApi,
 
   async querySourceFile(query: {
     fileId: string;
@@ -160,7 +43,7 @@ export const api = {
       await getTranslationStatus({
         path: { fileId },
         query: queryParams,
-        client,
+        client: getClient(),
       })
     );
     return {
@@ -198,7 +81,7 @@ export const api = {
             ? resolveCanonicalLocale(locale, customMapping)
             : undefined,
         },
-        client,
+        client: getClient(),
       })
     );
     // The single-file response omits fileFormat; Sanity downloads serialized
@@ -206,87 +89,18 @@ export const api = {
     return decodeBase64(response.data);
   },
 
-  async downloadFileBatch(files: DownloadFilesData['body']) {
-    if (files.length === 0) return { files: [], count: 0 };
-
-    const request = async (batch: DownloadFilesData['body']) =>
-      unwrapApiResult(
-        await downloadFiles({
-          body: batch.map((file) => ({
-            ...file,
-            locale: file.locale
-              ? resolveCanonicalLocale(file.locale, customMapping)
-              : undefined,
-          })),
-          client,
-        })
-      );
-    const responses = await processBatches(files, async (batch) => [
-      await request(batch),
-    ]);
-    return {
-      files: responses.flatMap((response) =>
-        response.files.map((file) => ({
-          ...file,
-          ...(file.locale && {
-            locale: resolveAliasLocale(file.locale, customMapping),
-          }),
-          data: decodeFileContent(file.data, file.fileFormat),
-          // OpenAPI currently types metadata as an open object; the API emits
-          // JSON values here, matching DownloadedFile's public contract.
-          metadata: file.metadata as DownloadedFile['metadata'],
-        }))
-      ),
-      count: responses.reduce((count, response) => count + response.count, 0),
-    };
-  },
-
-  async setupProject(
-    files: GenerateProjectContextData['body']['files'],
-    options: Omit<GenerateProjectContextData['body'], 'files'> = {}
+  async downloadFileBatch(
+    files: Parameters<typeof sharedApi.downloadFileBatch>[0]
   ) {
-    return unwrapApiResult(
-      await generateProjectContext({ body: { files, ...options }, client })
-    );
+    const { pending: _pending, ...result } =
+      await sharedApi.downloadFileBatch(files);
+    return result;
   },
 
   async awaitJobs(
     jobIds: readonly string[],
     options?: Parameters<typeof awaitApiJobs>[2]
   ) {
-    return awaitApiJobs(client, jobIds, options);
-  },
-
-  async queryFileData(body: GetFileInfoData['body']) {
-    const result = unwrapApiResult(
-      await getFileInfo({
-        body: {
-          ...body,
-          translatedFiles: body.translatedFiles?.map((file) => ({
-            ...file,
-            locale: resolveCanonicalLocale(file.locale, customMapping),
-          })),
-        },
-        client,
-      })
-    );
-    return {
-      ...result,
-      translatedFiles: result.translatedFiles.map((file) => ({
-        ...file,
-        locale: resolveAliasLocale(file.locale, customMapping),
-      })),
-      sourceFiles: result.sourceFiles.map((file) => ({
-        ...file,
-        sourceLocale: resolveAliasLocale(file.sourceLocale, customMapping),
-        locales: file.locales.map((locale) =>
-          resolveAliasLocale(locale, customMapping)
-        ),
-      })),
-    };
-  },
-
-  async createBranch(body: CreateBranchData['body']) {
-    return unwrapApiResult(await createBranch({ body, client }));
+    return awaitApiJobs(getClient(), jobIds, options);
   },
 };
